+++ ci/bootstrap.sh
@@ -21,31 +21,14 @@ cd "$(dirname "$0")/../"
 
 echo "--- Bootstrapping :boot:"
 
-SHARED_CI_DIR="$(pwd)/.shared-ci"
-CLONE_URL="git@github.com:spatialos/gdk-for-unity-shared-ci.git"
-PINNED_SHARED_CI_BRANCH=$(cat ./ci/shared-ci.pinned | cut -d' ' -f 1)
-PINNED_SHARED_CI_VERSION=$(cat ./ci/shared-ci.pinned | cut -d' ' -f 2)
-
-# Clone the HEAD of the shared CI repo into ".shared-ci"
-
-if [[ -d "${SHARED_CI_DIR}" ]]; then
-    rm -rf "${SHARED_CI_DIR}"
-fi
-
-mkdir "${SHARED_CI_DIR}"
-
-# Workaround for being unable to clone a specific commit with depth of 1.
-pushd "${SHARED_CI_DIR}"
-    git init
-    git remote add origin "${CLONE_URL}"
-    git fetch --depth 20 origin "${PINNED_SHARED_CI_BRANCH}"
-    git checkout "${PINNED_SHARED_CI_VERSION}"
-popd
+./ci/get-shared-ci.sh
 
 # Clone the GDK for Unity repository
 
 CLONE_URI="git@github.com:spatialos/gdk-for-unity.git"
-TARGET_DIRECTORY="$(realpath $(pwd)/../gdk-for-unity)"
+pushd ../
+    TARGET_DIRECTORY="$(pwd)/gdk-for-unity"
+popd 
 PINNED_BRANCH=$(cat ./gdk.pinned | cut -d' ' -f 1)
 PINNED_VERSION=$(cat ./gdk.pinned | cut -d' ' -f 2)
 SKIP_GDK=false
@@ -93,6 +76,10 @@ EXTRA_ARGS=""
 
 if [[ -n ${BUILDKITE:-} ]]; then
     EXTRA_ARGS="--copy"
+fi
+
+if [[ "${BUILDKITE_AGENT_META_DATA_OS}" == "darwin" ]]; then
+    PATH="${PATH}:/usr/local/share/dotnet"
 fi
 
 dotnet run -p ./.shared-ci/tools/PackageSymLinker/PackageSymLinker.csproj -- \